+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * Fixed: Prompt to save before exit enabled now by default. ([#193](https://github.com/DamirPorobic/ksnip/issues/193))
 * Fixed: Configuration Window not translated. ([#186](https://github.com/DamirPorobic/ksnip/issues/186))
 * Fixed: ksnip opens anyway with -s option specified. ([#213](https://github.com/DamirPorobic/ksnip/issues/213))
+* Fixed: Open Image with full size window doesn't resize main window. ([#194](https://github.com/DamirPorobic/ksnip/issues/194))
 * Fixed kImageAnnotator: Unable to select number annotation when clicking on the number without background. ([#46](https://github.com/DamirPorobic/kImageAnnotator/issues/46))
 * Fixed kImageAnnotator: Ctrl Modifier stuck on second or third screenshot with Ctrl-N. ([#58](https://github.com/DamirPorobic/kImageAnnotator/issues/58))
 * Fixed kImageAnnotator: Undo/Redo is now disabled during crop and scale operation. ([#56](https://github.com/DamirPorobic/kImageAnnotator/issues/56))